requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
authlib>=1.3.1
httpx>=0.27.0
itsdangerous>=2.2.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
jinja2>=3.1.4
pydantic>=2.7.4
pydantic-settings>=2.3.4
sqlalchemy>=2.0.31
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.134.0
google-genai>=0.5.0
websockets>=12.0
groq>=0.9.0
